+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -3,10 +3,10 @@ package nextstep.jwp.controller;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.dto.LoginRequest;
 import nextstep.jwp.exception.UnauthorizedException;
-import nextstep.jwp.http.HttpCookie;
+import org.apache.coyote.HttpCookie;
 import nextstep.jwp.http.QueryStringConverter;
-import nextstep.jwp.http.Session;
-import nextstep.jwp.http.SessionManager;
+import org.apache.coyote.Session;
+import org.apache.coyote.SessionManager;
 import nextstep.jwp.model.User;
 import nextstep.jwp.support.View;
 import org.apache.catalina.support.Resource;
