+++ tomcat/src/test/java/nextstep/jwp/http/LoginInterceptorTest.java
@@ -1,9 +1,7 @@
 package nextstep.jwp.http;
 
 import nextstep.jwp.interceptor.LoginInterceptor;
-import org.apache.coyote.Headers;
-import org.apache.coyote.HttpHeader;
-import org.apache.coyote.HttpMethod;
+import org.apache.coyote.*;
 import org.apache.coyote.support.Request;
 import org.apache.coyote.support.RequestInfo;
 import org.apache.coyote.support.Response;
